Move to sanitizer_dl
