+++ compiler-rt/lib/sanitizer_common/sanitizer_dl.h
@@ -21,6 +21,10 @@ namespace __sanitizer {
 // - the main program itself, that contains the sanitizer.
 const char* DladdrSelfFName(void);
 
+// Returns the base address of the ELF header, taking custom base offsets
+// into account.
+char* DladdrElfHeaderBase(void* ld, char* addr);
+
 }  // namespace __sanitizer
 
 #endif  // SANITIZER_DL_H
