+++ compiler-rt/lib/sanitizer_common/sanitizer_linux_libcdep.cpp
@@ -1094,53 +1094,6 @@ void InitializePlatformCommonFlags(CommonFlags *cf) {
 #  endif
 }
 
-#  if !SANITIZER_ANDROID && !SANITIZER_HAIKU
-// Call cb for each region mapped by map.
-void ForEachMappedRegion(link_map *map, void (*cb)(const void *, uptr)) {
-  CHECK_NE(map, nullptr);
-#    if !SANITIZER_FREEBSD && !SANITIZER_HAIKU
-  typedef ElfW(Phdr) Elf_Phdr;
-  typedef ElfW(Ehdr) Elf_Ehdr;
-#    endif  // !SANITIZER_FREEBSD
-  char *base = (char *)map->l_addr;
-#    if SANITIZER_GLIBC
-  Dl_info info;
-  if (dladdr((void *)map->l_ld, &info) && info.dli_fbase) {
-    base = (char *)info.dli_fbase;
-  }
-#    endif  // SANITIZER_GLIBC
-  Elf_Ehdr *ehdr = (Elf_Ehdr *)base;
-  char *phdrs = base + ehdr->e_phoff;
-  char *phdrs_end = phdrs + ehdr->e_phnum * ehdr->e_phentsize;
-
-  // Find the segment with the minimum base so we can "relocate" the p_vaddr
-  // fields.  Typically ET_DYN objects (DSOs) have base of zero and ET_EXEC
-  // objects have a non-zero base.
-  uptr preferred_base = (uptr)-1;
-  for (char *iter = phdrs; iter != phdrs_end; iter += ehdr->e_phentsize) {
-    Elf_Phdr *phdr = (Elf_Phdr *)iter;
-    if (phdr->p_type == PT_LOAD && preferred_base > (uptr)phdr->p_vaddr)
-      preferred_base = (uptr)phdr->p_vaddr;
-  }
-
-  // Compute the delta from the real base to get a relocation delta.
-  sptr delta = (uptr)base - preferred_base;
-  // Now we can figure out what the loader really mapped.
-  for (char *iter = phdrs; iter != phdrs_end; iter += ehdr->e_phentsize) {
-    Elf_Phdr *phdr = (Elf_Phdr *)iter;
-    if (phdr->p_type == PT_LOAD) {
-      uptr seg_start = phdr->p_vaddr + delta;
-      uptr seg_end = seg_start + phdr->p_memsz;
-      // None of these values are aligned.  We consider the ragged edges of the
-      // load command as defined, since they are mapped from the file.
-      seg_start = RoundDownTo(seg_start, GetPageSizeCached());
-      seg_end = RoundUpTo(seg_end, GetPageSizeCached());
-      cb((void *)seg_start, seg_end - seg_start);
-    }
-  }
-}
-#  endif
-
 }  // namespace __sanitizer
 
 #endif
